Add Update Function
- add update function
- change all id to integer

+++ src/main/java/com/product/repository/ProductRepository.java
@@ -1,5 +1,6 @@
 package com.product.repository;
 
+import com.product.dto.ProductSearchResultDTO;
 import com.product.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -9,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 
     @Query(value = "SELECT * FROM product_search(:query, :numberOfResults)", nativeQuery = true)
     List<Object[]> searchProducts(@Param("query") String searchQuery, @Param("numberOfResults") int numResults);
